// movies-store-mobile/src/app/_helpers/enums/table-names.ts
export enum TableNames {
  CARGO = 'cargo',
  CARGO_CONTRATISTA = 'cargo_contratista',
  GERENCIA = 'gerencia',
  MODULO = 'modulo',
  EMPRESA = 'empresa',
  UNIDAD_NEGOCIO = 'unidad_negocio',
  AREA_TRABAJO = 'area_trabajo',
  YACIMIENTO = 'yacimiento',
  LUGAR_ESPECIFICO = 'lugar_especifico',
  TIPO_USUARIO = 'tipo_usuario',
  CHECKLIST_CABECERA = 'checklist_cabecera',
  CHECKLIST_CABECERA_GRUPO = 'checklist_cabecera_grupo',
  CHECKLIST_CABECERA_GRUPO_ITEM = 'checklist_cabecera_grupo_item',
  INSPECCION_TOP = 'inspeccion_top',
  INSPECCION_TOP_IMAGEN = 'inspeccion_top_imagen'
}